requirements.txt
Flask>=2.3,<3
gunicorn>=20.1
google-cloud-firestore>=2.11
firebase-admin>=7.0.0
python-dotenv>=1.0
pydantic>=2.5
langchain>=0.1.0
langchain-core>=0.0.0
langchain-openai>=0.1.0
langgraph>=0.17
tavily>=0.1

